fix(komga): prevent panic on series list with readStatus condition filter
The POST /komga/api/v1/series/list endpoint panicked with a
TryFromIntError when the request included a readStatus condition
(e.g., from Komic). The readStatus filter path used u64::MAX as
page_size to fetch all series for in-memory filtering, but
sea-query-binder converts BigUnsigned(u64) to i64 via TryFrom for
both SQLite and PostgreSQL backends, which overflows on u64::MAX.

Replace u64::MAX with i64::MAX, matching the pattern already used in
the v1 books handler for the same reason.

+++ src/api/routes/komga/handlers/series.rs
@@ -314,7 +314,7 @@ async fn search_series_with_read_status_filter(
         search: search_term,
         sort: parse_komga_series_sort_param(sort_param),
         page: 0,
-        page_size: u64::MAX, // Fetch all for filtering
+        page_size: i64::MAX as u64, // Fetch all for filtering (i64::MAX avoids SQLite/PG integer overflow)
     };
 
     let (series_list, _) = SeriesRepository::query(&state.db, options)
